Tier1 Enemies (Green and Yellow)
Sorry about the lack of updates from me. I was in a real bad headspace and I couldn't focus on the project because I had to catch up on my grades, but here I am with the first 2 enemies. The hardest one was the yellow one because of how much time it took trying to figure out how to put together the patrolling/aggro system. The green one was pretty easy since I had a base for it, but I cut majority of what I had to complete it. But here I am with the first two enemies. Orange and Blue shouldn't be too difficult to do when I work on them. Anyways, see y'all this meeting.