Orange and Blue Enemies
This is mainly for the Orange and Blue enemies. The Blue one took me a long time to get right, the Orange one kept stopping on a flat surface for some reason, and I had to make an enemy layer to make sure that the layers collide with the right layers. Still couldn't make the hitboxes, but I think that setting all the hitboxes to "damagePlayer = true" in their Scriptable Objects can make do. Anyways, time to make do with what we have.