main -> Demo (#5)
* Tier1 Enemies (Green and Yellow)

Sorry about the lack of updates from me. I was in a real bad headspace and I couldn't focus on the project because I had to catch up on my grades, but here I am with the first 2 enemies. The hardest one was the yellow one because of how much time it took trying to figure out how to put together the patrolling/aggro system. The green one was pretty easy since I had a base for it, but I cut majority of what I had to complete it. But here I am with the first two enemies. Orange and Blue shouldn't be too difficult to do when I work on them. Anyways, see y'all this meeting.

* Orange and Blue Enemies

This is mainly for the Orange and Blue enemies. The Blue one took me a long time to get right, the Orange one kept stopping on a flat surface for some reason, and I had to make an enemy layer to make sure that the layers collide with the right layers. Still couldn't make the hitboxes, but I think that setting all the hitboxes to "damagePlayer = true" in their Scriptable Objects can make do. Anyways, time to make do with what we have.

* Trying to fit in my pull request to main

Yeah, I just had to delete one thing before I merged